# backend/src/dtos/user_dto.ts
export interface UserDto {
  _id: string | undefined;
  fullName: string | undefined;
  userName: string | undefined;
  age: number | undefined;
  gender: string | undefined;
  phone: string | undefined;
  email: string | undefined;
  city: string | undefined;
}